# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Collaborative coding platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserUpdate {
  name?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  language: string;
  ownerId: number;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  ownerAvatarUrl?: string | null;
  isPublic?: boolean;
  memberCount?: number;
  fileCount?: number;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  lastOpenedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  language: string;
  isPublic?: boolean;
}

export interface ProjectUpdate {
  name?: string;
  description?: string;
  language?: string;
  isPublic?: boolean;
}

export type ProjectMemberRole = typeof ProjectMemberRole[keyof typeof ProjectMemberRole];


export const ProjectMemberRole = {
  owner: 'owner',
  admin: 'admin',
  editor: 'editor',
  viewer: 'viewer',
} as const;

export interface ProjectMember {
  userId: number;
  projectId: number;
  role: ProjectMemberRole;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  joinedAt: string;
}

export type InviteInputRole = typeof InviteInputRole[keyof typeof InviteInputRole];


export const InviteInputRole = {
  admin: 'admin',
  editor: 'editor',
  viewer: 'viewer',
} as const;

export interface InviteInput {
  email: string;
  role: InviteInputRole;
}

export type RoleUpdateRole = typeof RoleUpdateRole[keyof typeof RoleUpdateRole];


export const RoleUpdateRole = {
  admin: 'admin',
  editor: 'editor',
  viewer: 'viewer',
} as const;

export interface RoleUpdate {
  role: RoleUpdateRole;
}

export interface WeeklyActivityPoint {
  day: string;
  edits: number;
  messages: number;
}

export interface ProjectStats {
  totalEdits: number;
  totalCommits: number;
  activeUsers: number;
  filesModifiedToday: number;
  chatActivity: number;
  /** @nullable */
  mostActiveUser?: string | null;
  /** @nullable */
  mostActiveUserAvatarUrl?: string | null;
  weeklyActivity: WeeklyActivityPoint[];
}

export type ProjectFileType = typeof ProjectFileType[keyof typeof ProjectFileType];


export const ProjectFileType = {
  file: 'file',
  folder: 'folder',
} as const;

export interface ProjectFile {
  id: number;
  projectId: number;
  name: string;
  path: string;
  type: ProjectFileType;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  parentId?: number | null;
  /** @nullable */
  createdById?: number | null;
  /** @nullable */
  createdByName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type FileInputType = typeof FileInputType[keyof typeof FileInputType];


export const FileInputType = {
  file: 'file',
  folder: 'folder',
} as const;

export interface FileInput {
  /** @minLength 1 */
  name: string;
  type: FileInputType;
  /** @nullable */
  parentId?: number | null;
  language?: string;
  content?: string;
}

export interface FileUpdate {
  name?: string;
  content?: string;
  language?: string;
}

export interface FileVersion {
  id: number;
  fileId: number;
  content: string;
  /** @nullable */
  authorId?: number | null;
  /** @nullable */
  authorName?: string | null;
  createdAt: string;
}

export interface ActivityLog {
  id: number;
  projectId: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userAvatarUrl?: string | null;
  action: string;
  /** @nullable */
  targetType?: string | null;
  /** @nullable */
  targetName?: string | null;
  /** @nullable */
  metadata?: string | null;
  createdAt: string;
}

export type ChatMessageType = typeof ChatMessageType[keyof typeof ChatMessageType];


export const ChatMessageType = {
  text: 'text',
  code: 'code',
  file: 'file',
} as const;

export interface ChatMessage {
  id: number;
  projectId: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userAvatarUrl?: string | null;
  content: string;
  type?: ChatMessageType;
  /** @nullable */
  replyToId?: number | null;
  isEdited?: boolean;
  createdAt: string;
  updatedAt?: string;
}

export type MessageInputType = typeof MessageInputType[keyof typeof MessageInputType];


export const MessageInputType = {
  text: 'text',
  code: 'code',
  file: 'file',
} as const;

export interface MessageInput {
  /** @minLength 1 */
  content: string;
  type?: MessageInputType;
  /** @nullable */
  replyToId?: number | null;
}

export interface MessageUpdate {
  /** @minLength 1 */
  content: string;
}

export interface Notification {
  id: number;
  userId: number;
  type: string;
  message: string;
  isRead: boolean;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  fromUserId?: number | null;
  /** @nullable */
  fromUserName?: string | null;
  createdAt: string;
}

export interface DashboardSummary {
  totalProjects: number;
  ownedProjects: number;
  sharedProjects: number;
  recentActivity: ActivityLog[];
  unreadNotifications: number;
  recentProjects?: Project[];
}

export type ListProjectsParams = {
filter?: ListProjectsFilter;
search?: string;
};

export type ListProjectsFilter = typeof ListProjectsFilter[keyof typeof ListProjectsFilter];


export const ListProjectsFilter = {
  all: 'all',
  owned: 'owned',
  shared: 'shared',
  recent: 'recent',
} as const;

export type ListNotificationsParams = {
unread?: boolean;
};