# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Collaborative coding platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # ── Users ───────────────────────────────────────────────────────
...
  /projects/{projectId}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List all files in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectFile"

    post:
      operationId: createFile
      tags: [files]
      summary: Create a file or folder
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
  /projects/{projectId}/files/{fileId}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get file content
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
        "404":
          description: Not found

    patch:
      operationId: updateFile
      tags: [files]
      summary: Update file name or content
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"

    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/files/{fileId}/versions:
    get:
      operationId: listFileVersions
      tags: [files]
      summary: List file version history
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileVersion"
  /projects/{projectId}/files/{fileId}/versions/{versionId}/restore:
    post:
      operationId: restoreFileVersion
      tags: [files]
      summary: Restore a file to a previous version
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"

  # ── Activity ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        avatarUrl:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ── Users ─────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    ProjectFile:
      type: object
      required: [id, projectId, name, path, type, createdAt, updatedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, folder]
        language:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        parentId:
          type: ["integer", "null"]
        createdById:
          type: ["integer", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FileInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [file, folder]
        parentId:
          type: ["integer", "null"]
        language:
          type: string
        content:
          type: string
    FileUpdate:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
        language:
          type: string
    FileVersion:
      type: object
      required: [id, fileId, content, createdAt]
      properties:
        id:
          type: integer
        fileId:
          type: integer
        content:
          type: string
        authorId:
          type: ["integer", "null"]
        authorName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ── Activity ──────────────────────────────────────────────────
...